app/requirements.txt
flask
requests
pandas
matplotlib
plotly
python-dotenv